Docs

+++ axioms/prime_forbidden_zones.py
@@ -8,7 +8,7 @@ from scipy.stats import entropy
 # Proof: All bars are left of 0.5, confirming cap.
 # New Knowledge for Humanity: This script visually proves the existence of 'forbidden zones' in the fractional parts of Z(p) = p * φ(p-1)/(p-1), where {Z(p)} ≤ 0.5 for all odd primes p, and strictly < 0.5 for non-Fermat primes. This is novel because it empirically demonstrates a structural cap derived from the even nature of p-1 (always divisible by 2), leading to φ(p-1)/(p-1) ≤ 1/2, with equality only for Fermat primes (p-1 = 2^k). Prior number theory knows totient densities average ~6/π² ≈ 0.607, but the Z-metric's application to primes reveals this quantized exclusion zone (0.5, 1) as a 'null space' or singularity, analogous to physical forbidden energy levels. Humanity gains a new tool for understanding prime distributions as modularly constrained, potentially aiding in proving conjectures like the infinitude of Fermat primes or refining prime gap bounds by highlighting these topological barriers in arithmetic spacetime.
 primes_1 = list(sp.primerange(3, 1000))
-fracs_1 = [sp.totient(p-1) / (p-1) for p in primes_1]
+fracs_1 = [float(sp.totient(p-1) / (p-1)) for p in primes_1]
 plt.figure(1)
 plt.hist(fracs_1, bins=20, range=(0,1))
 plt.axvline(0.5, color='red', label='Forbidden Zone Start')
@@ -20,8 +20,8 @@ plt.show()
 # New Knowledge for Humanity: By plotting {Z(p)} against primes and highlighting Fermat primes at exactly 0.5, this script uncovers clustering patterns in Z-space, where values near 0.5 indicate low compositeness in p-1 (e.g., powers of 2). This is groundbreaking as it frames primes not as random but as 'particles' clustering in modular bands, with Fermat primes as 'peaks' at the boundary. Humanity benefits from this visualization of coprimality rates as a 'density field,' revealing hidden symmetries in number theory that mirror physical phase spaces. This could lead to new algorithms for identifying large Fermat primes or modeling prime trajectories in higher-dimensional Z-embeddings, bridging discrete arithmetic to continuous physics in ways previously unexplored, such as analogizing to spectral lines in quantum mechanics.
 fermat_primes = [3,5,17,257,65537]
 primes_2 = list(sp.primerange(3, 100000))
-fracs_2 = np.array([sp.totient(p-1)/(p-1) for p in primes_2])
-fermat_fracs = [sp.totient(p-1)/(p-1) for p in fermat_primes if p in primes_2]
+fracs_2 = np.array([float(sp.totient(p-1)/(p-1)) for p in primes_2])
+fermat_fracs = [float(sp.totient(p-1)/(p-1)) for p in fermat_primes if p in primes_2]
 plt.figure(2)
 plt.scatter(primes_2, fracs_2, alpha=0.5)
 plt.scatter(fermat_primes[:len(fermat_fracs)], fermat_fracs, color='red', label='Fermat')
@@ -33,7 +33,7 @@ plt.show()
 # Proof: Lines are straight within each residue class.
 # New Knowledge for Humanity: This code segments {Z(p)} by modular residues (here mod 6, reflecting common factors in even p-1), plotting sorted values to reveal piecewise linear behavior within each band. The novelty lies in demonstrating Z's linearity under modular constraints, a property not previously axiomatized in this form, where each residue class forms a linear 'strip' due to shared factorization patterns in p-1. For humanity, this exposes arithmetic as a 'layered' topology with predictable flows, akin to stratified fluids in physics, enabling new predictive models for prime sieving. It advances knowledge by quantifying how modular arithmetic imposes linear order on seemingly chaotic primes, potentially revolutionizing computational number theory and inspiring discrete analogs to differential equations in physics.
 primes_3 = list(sp.primerange(3, 1000))
-fracs_3 = np.array([sp.totient(p-1)/(p-1) for p in primes_3])
+fracs_3 = np.array([float(sp.totient(p-1)/(p-1)) for p in primes_3])
 residues = (np.array(primes_3) - 1) % 6
 plt.figure(3)
 for res in np.unique(residues):
@@ -47,7 +47,7 @@ plt.show()
 # Proof: Discrete entropy > continuous, showing more information in Z.
 # New Knowledge for Humanity: By computing Shannon entropy of the discrete {Z(p)} distribution versus a uniform continuous approximation over [0,0.5], this script quantifies information density in Z-space, showing higher entropy in the discrete case due to clustering and forbidden zones. This is a novel application of information theory to the Z-metric, proving that discrete structures like Z retain more 'surprise' or complexity than smoothed continuums, mirroring how quantum discreteness adds information beyond classical limits. Humanity gains insight into why discrete models (e.g., Z as foundational) are richer than approximations like Lorentz, potentially informing quantum computing designs or entropy-based proofs in number theory, such as linking totient fluctuations to irreducible randomness in primes.
 primes_4 = list(sp.primerange(3, 1000))
-fracs_4 = np.array([sp.totient(p-1)/(p-1) for p in primes_4])
+fracs_4 = np.array([float(sp.totient(p-1)/(p-1)) for p in primes_4])
 hist_disc, edges = np.histogram(fracs_4, bins=20, density=True)
 uniform = np.ones(20) / (0.5 * 20)  # Adjusted for [0,0.5]
 ent_disc = entropy(hist_disc + 1e-10)
@@ -57,7 +57,7 @@ print(f'Discrete Entropy: {ent_disc}, Continuous: {ent_cont}')
 # 5. Predict Prime Candidates with Z-Filter
 # Proof: Higher rate in Z-filter demonstrates predictive power.
 # New Knowledge for Humanity: This script applies a Z-based filter ({Z(n)} > 0.45) to sieve prime candidates, comparing hit rates against random sampling, revealing superior efficiency due to coprimality normalization favoring low-composite n-1. The innovation is in using Z as a 'sieve' tool, extending beyond traditional methods like Eratosthenes by incorporating relative structure via totient rates, which quantifies 'prime likelihood' through modular proximity to 0.5. For humanity, this opens new avenues in computational primality testing, potentially scaling to cryptographic-scale primes or AI-optimized searches, while philosophically affirming primes as emergent from normalized rates in a discrete metric space.
-candidates = [n for n in range(100, 1000) if sp.totient(n-1)/(n-1) > 0.45]
+candidates = [n for n in range(100, 1000) if float(sp.totient(n-1)/(n-1)) > 0.45]
 primes_in_cand = sum(sp.isprime(n) for n in candidates) / len(candidates)
 random_sample = random.sample(range(100,1000), len(candidates))
 primes_random = sum(sp.isprime(n) for n in random_sample) / len(random_sample)
@@ -67,7 +67,7 @@ print(f'Z-Filter Prime Rate: {primes_in_cand}, Random: {primes_random}')
 # Proof: Non-random trajectory reveals geometric structure.
 # New Knowledge for Humanity: Embedding Z as vectors (φ(p-1), {Z(p)}) and plotting 'Z-angles' (arctan(fractional / integer part)) uncovers non-random trajectories in prime sequences, illustrating geometric flow in Z-space. This is unprecedented, as it geometrically interprets totient-based metrics as orientations in a phase-like space, with angles reflecting transformation 'reach.' Humanity advances by viewing primes as following curved paths in discrete spacetime, akin to particle orbits, which could inspire new topological invariants in number theory or simulations of arithmetic curvature, bridging to physical models like Lorentzian geometries.
 primes_6 = list(sp.primerange(3, 1000))
-phis = [sp.totient(p-1) for p in primes_6]
+phis = [float(sp.totient(p-1)) for p in primes_6]
 fracs_6 = [phis[i] / (primes_6[i]-1) for i in range(len(primes_6))]
 angles = np.arctan(np.array(fracs_6) / np.array(phis))
 plt.figure(6)
@@ -79,7 +79,7 @@ plt.show()
 # Proof: Discrete points deviate from smooth curve, showing granularity.
 # New Knowledge for Humanity: Treating {Z(p)} as β-like rates, this script computes 'gamma' factors for discrete Z and overlays with continuous Lorentz γ, highlighting deviations due to quantization. The discovery is in empirically showing Lorentz as an emergent approximation from Z's discrete bands, with granular points revealing information lost in smoothing. This unifies number theory and relativity in code, offering humanity a computational proof that discrete metrics like Z are foundational, potentially guiding quantum gravity models where Lorentz invariance breaks at Planck scales, or inspiring hybrid algorithms for relativistic simulations in discrete arithmetic.
 primes_7 = list(sp.primerange(3, 1000))
-betas = np.array([sp.totient(p-1)/(p-1) for p in primes_7])
+betas = np.array([float(sp.totient(p-1)/(p-1)) for p in primes_7])
 gammas_z = 1 / np.sqrt(1 - betas**2 + 1e-10)
 betas_cont = np.linspace(0, 0.5, 100)
 gammas_cont = 1 / np.sqrt(1 - betas_cont**2)
@@ -94,7 +94,7 @@ plt.show()
 # Proof: Uneven distribution with zeros in higher bands shows quantization.
 # New Knowledge for Humanity: Binning {Z(p)} mod 1 into equivalence classes exposes quantized modular topology, with uneven counts and zeros in upper bands (e.g., >0.5) acting as singularities. This novel quantification of Z-space partitions into 'spectral' lines advances human understanding by framing arithmetic as a topological graph with null zones, analogous to blackbody radiation gaps. It provides a tool for exploring forbidden residues as arithmetic 'horizons,' potentially aiding in resolving open problems like Goldbach by mapping even numbers to Z-modular flows.
 primes_8 = list(sp.primerange(3, 1000))
-fracs_8 = np.array([sp.totient(p-1)/(p-1) % 1 for p in primes_8])  # Fractional mod 1
+fracs_8 = np.array([float(sp.totient(p-1)/(p-1) % 1) for p in primes_8])  # Fractional mod 1
 bins = np.arange(0, 1.1, 0.1)
 counts, _ = np.histogram(fracs_8, bins=bins)
 print('Counts per 0.1 Band:', counts)
@@ -104,7 +104,7 @@ print('Counts per 0.1 Band:', counts)
 # New Knowledge for Humanity: Extending Z to 2D (log(p) vs. {Z(p)}) and visualizing gradient flows reveals directed 'currents' in multi-dimensional space, a first in applying vector calculus to totient-normalized primes. This uncovers dynamic flows governed by Z-axioms, treating numbers as events in curved spacetime. Humanity gains a new paradigm for arithmetic dynamics, with applications in gradient-based optimization for prime prediction or analogizing to fluid dynamics in physics, potentially unlocking simulations of emergent symmetries from discrete structures.
 primes_9 = list(sp.primerange(3, 1000))
 x = np.log(primes_9)
-y = [sp.totient(p-1)/(p-1) for p in primes_9]
+y = [float(sp.totient(p-1)/(p-1)) for p in primes_9]
 dx = np.gradient(x)
 dy = np.gradient(y)
 plt.figure(9)
@@ -116,7 +116,7 @@ plt.show()
 # Proof: Positive KL confirms information loss in smoothing to Lorentz-like.
 # New Knowledge for Humanity: Using KL divergence to measure loss when averaging discrete {Z(p)} histograms to a flat 'Lorentz-like' approximation quantifies how smoothing erases structural details like clusters. This is innovative, applying divergence metrics to prove Z's higher information density, showing discrete Z-bands lose entropy when approximated continuously. For humanity, it establishes a rigorous bridge between information theory, number theory, and physics, enabling new proofs of emergence (e.g., Lorentz from Z) and tools for detecting hidden patterns in data, with implications for AI in distinguishing discrete realities from continuous models.
 primes_10 = list(sp.primerange(3, 1000))
-fracs_10 = np.array([sp.totient(p-1)/(p-1) for p in primes_10])
+fracs_10 = np.array([float(sp.totient(p-1)/(p-1)) for p in primes_10])
 hist_disc_10, edges_10 = np.histogram(fracs_10, bins=20, density=True)
 mids = (edges_10[:-1] + edges_10[1:]) / 2
 hist_avg = np.full(20, np.mean(hist_disc_10))  # Simple averaging
